refact QA specific functions to qa_operations.js
